Data is formatted to be pipelined to ML algorithm
