+++ .gitignore
@@ -130,3 +130,4 @@ dmypy.json
 
 .vscode/
 try.py
+example_data/
