+++ requirements.txt
@@ -1,5 +1,8 @@
 DateTime==4.3
 pandas==1.0.4
+plotly==4.8.1
 python-dateutil==2.8.1
 requests==2.23.0
 urllib3==1.25.9
+yfinance==0.1.54
+jupyter==1.0.0
